bigdata/requirements.txt
# Python Dependencies for BigData Pipeline
# Install with: pip install -r requirements.txt

# Kafka
kafka-python==2.0.2

# Spark
pyspark==3.5.0

# PostgreSQL
psycopg2-binary==2.9.9

# Monitoring & Metrics
prometheus-client==0.19.0

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# Logging
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality (optional)
black==23.12.1
flake8==6.1.0
mypy==1.7.1
